requirements.txt
torch==2.5.1
numpy==2.2.1
pettingzoo==1.24.3
rlcard==1.2.0
gymnasium==1.0.0
PyYAML==6.0.2
pygame==2.6.1
gradio==5.9.1
pillow==11.0.0
